# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API for Golang Project TODO.",
        "title": "TODO App Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API [Roman] Support"
        },
        "license": {
            "name": "romandnk",
            "url": "https://github.com/romandnk/todo"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
        "/statuses/": {
            "post": {
                "description": "Create new task status.",
                "tags": [
                    "Status"
                ],
                "summary": "CreateStatus",
                "parameters": [
                    {
                        "description": "Required JSON body with status name",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateStatusParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status was created successfully",
                        "schema": {
                            "$ref": "#/definitions/service.CreateStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/tasks/": {
            "post": {
                "description": "Create new task.",
                "tags": [
                    "Task"
                ],
                "summary": "CreateTask",
                "parameters": [
                    {
                        "description": "Required JSON body with all required task field",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateTaskParams"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task was created successfully",
                        "schema": {
                            "$ref": "#/definitions/service.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/tasks/:id": {
            "delete": {
                "description": "Delete task by its id.",
                "tags": [
                    "Task"
                ],
                "summary": "DeleteTaskByID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Required task id to delete",
                        "name": "params",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task was deleted successfully"
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.CreateStatusParams": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "service.CreateStatusResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "service.CreateTaskParams": {
            "type": "object",
            "required": [
                "date",
                "description",
                "status_name",
                "title"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "service.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}